+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- DSC for SharePoint
+  - Add `ProjectServerService` to `SharePointConfigurations`, to deploy Project Server
 - Customized Modules
   - SharePointDSC: Set ErrorAction to `Stop` instead of `SilentlyContinue` in `Write-EventLog`, which is the same change as https://github.com/dsccommunity/SharePointDsc/pull/1473